comentarii si api cont

+++ api/category/delete.php
@@ -1,5 +1,12 @@
 <?php
 //fisierul asta va sterge o categorie; id-ul categoriei va fi dat de pagina respectiva
+/*
+  Cum se foloseste api-ul asta: metoda POST, link-ul http://localhost/api/category/delete.php
+  si corpul json:
+  {
+   "id_categorie" : idCategorieDinObiectCRED
+  }
+*/
 //TODO incomplet, va trebui realizata stergerea tuturor conturilor din categoria respectiva
 header("Access-Control-Allow-Origin: *");
 header("Content-Type: application/json; charset=UTF-8");
@@ -19,8 +26,14 @@ $data = json_decode(file_get_contents("php://input"));
 
 $category->id_categorie = $data->id_categorie;
 
+if(!ctype_digit($category->id_categorie)) {
+  http_response_code(400);
+  echo json_encode(array("message" => "Invalid category id detected."));
+  die();
+}
+
 if($category->delete()){
-    http_response_code(200);
+    http_response_code(204);
     echo json_encode(array("message" => "Category was deleted."));
 } else {
     http_response_code(503);
